requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
python-dotenv==1.0.1
astrapy==1.4.1
flask==3.0.3
pypdf==4.3.1
docx2txt==0.8
openai==1.44.0
tiktoken==0.7.0
gunicorn==23.0.0
youtube-transcript-api==0.6.2
Flask-SocketIO==5.4.1
azure-cognitiveservices-speech==1.41.1
gevent==24.11.1
gevent-websocket==0.10.1
flask-cors